kv/storage/standalone_storage/standalone_storage.go
package standalone_storage

import (
	"github.com/Connor1996/badger"
	"github.com/pingcap-incubator/tinykv/kv/config"
	"github.com/pingcap-incubator/tinykv/kv/storage"
	"github.com/pingcap-incubator/tinykv/kv/util/engine_util"
	tinkvlog "github.com/pingcap-incubator/tinykv/log"
	"github.com/pingcap-incubator/tinykv/proto/pkg/kvrpcpb"
	"log"
)

// StandAloneStorage is an implementation of `Storage` for a single-node TinyKV instance. It does not
// communicate with other nodes and all data is stored locally.
type StandAloneStorage struct {
	// Your Data Here (1).
	kvDB *badger.DB
	dbPath string
}

func NewStandAloneStorage(conf *config.Config) *StandAloneStorage {
	// Your Code Here (1).
	return &StandAloneStorage{
		kvDB: nil,
		dbPath: conf.DBPath,
	}
}

func (s *StandAloneStorage) Start() error {
	// Your Code Here (1).
	opt := badger.DefaultOptions
	opt.Dir = s.dbPath
	opt.ValueDir = s.dbPath
	db, err := badger.Open(opt)
	if err != nil {
		tinkvlog.Error(err)
		return err
	}
	s.kvDB = db
	return nil
}

func (s *StandAloneStorage) Stop() error {
	// Your Code Here (1).
	err := s.kvDB.Close()
	if err != nil {
		tinkvlog.Error(err)
		return err
	}
	return nil
}

func (s *StandAloneStorage) Reader(ctx *kvrpcpb.Context) (storage.StorageReader, error) {
	// Your Code Here (1).
	storageReader, err := NewStandAloneStorageReader(s)
	if err != nil {
		return nil, err
	}
	return storageReader, err
}

func (s *StandAloneStorage) Write(ctx *kvrpcpb.Context, batch []storage.Modify) error {
	// Your Code Here (1).
	wb := new(engine_util.WriteBatch)
	for _, v := range batch {
		wb.SetCF(v.Cf(), v.Key(), v.Value())
	}
	err := wb.WriteToDB(s.kvDB)
	if err != nil {
		log.Println(err)
		return err
	}
	return nil
}